Moved datasets to separate files.
